+++ apps/chatbot/src/modules/monitor.py
@@ -29,12 +29,13 @@ def get_trace(trace_id: str, as_dict: bool = False) -> TraceWithFullDetails | di
         TraceWithFullDetails | dict: The trace object or its dictionary representation.
     """
 
+    trace = None
     LOGGER.info(f"Getting trace {trace_id} from Langfuse")
     try:
         trace = LANGFUSE_CLIENT.fetch_trace(trace_id)
         trace = trace.data
     except Exception as e:
-        LOGGER.error(e)
+        LOGGER.error(f"LANGFUSE_CLIENT.fetch_trace: {e}")
 
     if as_dict:
         return trace.dict()
@@ -101,7 +102,7 @@ def add_langfuse_score(
     """
 
     if value is None or (isinstance(value, float) and np.isnan(value)):
-        LOGGER.warning(
+        LOGGER.info(
             f"Value for score {name} is None or NaN, setting to 0.0 for trace {trace_id}."
         )
         value = 0.0
@@ -111,11 +112,13 @@ def add_langfuse_score(
     found_score = False
     if score_id is None:
         trace = get_trace(trace_id)
-        for score in trace.scores:
-            if score.name == name:
-                score_id = score.id
-                found_score = True
-                break
+        
+        if trace is not None:
+            for score in trace.scores:
+                if score.name == name:
+                    score_id = score.id
+                    found_score = True
+                    break
 
         if not found_score:
             score_id = str(uuid.uuid4())
@@ -134,7 +137,7 @@ def add_langfuse_score(
         LANGFUSE_CLIENT.flush()
 
         if found_score:
-            LOGGER.info(f"Updated {name} score with value {value} in trace {trace_id}.")
+            LOGGER.warning(f"Updated {name} score with value {value} in trace {trace_id}.")
         else:
             LOGGER.info(f"Added {name} score with value {value} in trace {trace_id}.")
     except Exception as e: