+++ apps/nextjs-website/src/components/organisms/WebinarsTemplate/WebinarsTemplate.tsx
@@ -2,9 +2,12 @@
 import React, { useEffect, useState, Suspense } from 'react';
 import Hero from '@/editorialComponents/Hero/Hero';
 import { useTranslations } from 'next-intl';
-import { useTheme } from '@mui/material';
+import { Box, useTheme } from '@mui/material';
 import { Webinar } from '@/lib/types/webinar';
-import { getFutureWebinars, getPastWebinars } from '@/helpers/webinars.helpers';
+import {
+  getFutureWebinarsFrom,
+  getPastWebinarsFrom,
+} from '@/helpers/webinars.helpers';
 import FutureWebinarsShowcase from '../FutureWebinarsShowcase/FutureWebinarsShowcase';
 import { baseUrl } from '@/config';
 import { generateStructuredDataScripts } from '@/helpers/generateStructuredDataScripts.helpers';
@@ -12,6 +15,7 @@ import { getItemFromPaths } from '@/helpers/structuredData.helpers';
 import Spinner from '@/components/atoms/Spinner/Spinner';
 import { FilteredGridLayout } from '@/components/organisms/FilteredGridLayout/FilteredGridLayout';
 import { Tag } from '@/lib/types/tag';
+import { SITE_HEADER_HEIGHT } from '@/config';
 
 const CHECK_WEBINARS_INTERVAL_MS = 60 * 1000;
 
@@ -45,18 +49,18 @@ const WebinarsTemplateContent = ({
     seo: webinarsListPageSEO,
   });
   useEffect(() => {
-    setFutureWebinars(getFutureWebinars(webinars));
-    setPastWebinars(getPastWebinars(webinars));
+    setFutureWebinars(getFutureWebinarsFrom(webinars));
+    setPastWebinars(getPastWebinarsFrom(webinars));
 
     const intervalId = setInterval(() => {
-      setFutureWebinars(getFutureWebinars(webinars));
-      setPastWebinars(getPastWebinars(webinars));
+      setFutureWebinars(getFutureWebinarsFrom(webinars));
+      setPastWebinars(getPastWebinarsFrom(webinars));
     }, CHECK_WEBINARS_INTERVAL_MS);
 
     // Cleanup the interval when the component is unmounted
     return () => clearInterval(intervalId);
   }, [webinars]);
-  const mappedWebinars = webinars.map((webinar) => ({
+  const mappedWebinars = pastWebinars.map((webinar) => ({
     tags: webinar.tag ? [webinar.tag] : [],
     title: webinar.title,
     date: {
@@ -101,9 +105,15 @@ const WebinarsTemplateContent = ({
 
 const WebinarsTemplate = (props: WebinarsTemplateProps) => {
   return (
-    <Suspense fallback={<Spinner />}>
-      <WebinarsTemplateContent {...props} />
-    </Suspense>
+    <Box
+      sx={{
+        marginTop: `${SITE_HEADER_HEIGHT}px`,
+      }}
+    >
+      <Suspense fallback={<Spinner />}>
+        <WebinarsTemplateContent {...props} />
+      </Suspense>
+    </Box>
   );
 };