+++ apps/nextjs-website/src/components/templates/GitBookTemplate/GitBookTemplate.tsx
@@ -8,6 +8,7 @@ import GuideInPageMenu from '@/components/organisms/GuideInPageMenu/GuideInPageM
 import { BreadcrumbSegment } from '@/lib/types/path';
 import { Box, Stack } from '@mui/material';
 import { useTranslations } from 'next-intl';
+import { PRODUCT_HEADER_HEIGHT, SITE_HEADER_HEIGHT } from '@/config';
 import GitBookContent from '@/components/organisms/GitBookContent/GitBookContent';
 
 export type GitBookTemplateProps = {
@@ -19,6 +20,7 @@ export type GitBookTemplateProps = {
   versionName?: GuideMenuItemsProps['versionName'];
   hasHeader?: boolean;
   hasInPageMenu?: boolean;
+  hasProductHeader?: boolean;
 } & Pick<
   ProductGuidePageProps,
   'menu' | 'body' | 'bodyConfig' | 'path' | 'pathPrefix'
@@ -35,12 +37,14 @@ const GitBookTemplate = ({
   versions,
   breadcrumbs,
   menuDistanceFromTop,
-  contentMarginTop = 75,
+  contentMarginTop,
   hasHeader = true,
   hasInPageMenu = true,
+  hasProductHeader = true,
 }: GitBookTemplateProps) => {
   const t = useTranslations();
-  const paddingTop = hasHeader ? '60px' : '-80px';
+  const responsiveContentMarginTop =
+    (contentMarginTop && `${contentMarginTop}px`) || 0;
 
   return (
     <FragmentProvider>
@@ -62,32 +66,34 @@ const GitBookTemplate = ({
             versionName={versionName}
             versions={versions}
             distanceFromTop={menuDistanceFromTop}
+            hasProductHeader={hasProductHeader}
           />
         )}
         <Stack
           sx={{
-            margin: `${contentMarginTop} auto`,
-            paddingTop: 3,
+            marginTop: responsiveContentMarginTop,
+            paddingTop: { xs: '68px', sm: '78px', md: '24px' },
             flexGrow: { lg: 1 },
             maxWidth: {
               xs: '100%',
               lg: '1008px',
             },
+            minHeight: '100vh',
           }}
         >
-          <Box sx={{ paddingTop: paddingTop, paddingX: '40px' }}>
+          <Box sx={{ paddingX: '40px' }}>
             <ProductBreadcrumbs breadcrumbs={breadcrumbs} />
           </Box>
-          <Box sx={{ padding: '32px 40px' }}>
+          <Box sx={{ padding: '0 40px 32px 40px' }}>
             <GitBookContent content={body} config={bodyConfig} />
           </Box>
         </Stack>
         {hasInPageMenu && (
           <Box
             sx={{
-              display: { xs: 'none', lg: 'initial' },
+              display: { lg: 'none', xl: 'initial' },
               position: 'relative',
-              padding: { lg: hasHeader ? '80px 64px' : '48px 64px' },
+              padding: { lg: '0 32px', xl: '0 64px 0 32px' },
               width: { lg: '378px' },
             }}
           >
@@ -95,7 +101,9 @@ const GitBookTemplate = ({
               sx={{
                 position: 'sticky',
                 maxWidth: '378px',
-                top: hasHeader ? 144 : 64,
+                top:
+                  SITE_HEADER_HEIGHT +
+                  (hasProductHeader ? PRODUCT_HEADER_HEIGHT : 0),
               }}
             >
               <GuideInPageMenu
@@ -103,6 +111,7 @@ const GitBookTemplate = ({
                 pagePath={path}
                 inPageMenu={body}
                 title={t('productGuidePage.onThisPage')}
+                hasProductHeader={hasProductHeader}
               />
             </Box>
           </Box>