+++ apps/chatbot-index/CHANGELOG.md
@@ -1,5 +1,23 @@
 # chatbot-index
 
+## 2.1.0
+
+### Minor Changes
+
+- b6291ed: Add Docker compose for test
+
+## 2.0.2
+
+### Patch Changes
+
+- 07d1049: Align settings.py file to chatbot module
+
+## 2.0.1
+
+### Patch Changes
+
+- 612d7d7: Fix chrome install
+
 ## 2.0.0
 
 ### Major Changes